+++ src/components/dialog/dialog.ts
@@ -7,6 +7,7 @@ import { WaHideEvent } from '../../events/hide.js';
 import { WaShowEvent } from '../../events/show.js';
 import { animateWithClass } from '../../internal/animate.js';
 import { lockBodyScrolling, unlockBodyScrolling } from '../../internal/scroll.js';
+import { HasSlotController } from '../../internal/slot.js';
 import { watch } from '../../internal/watch.js';
 import WebAwesomeElement from '../../internal/webawesome-element.js';
 import { LocalizeController } from '../../utilities/localize.js';
@@ -56,6 +57,7 @@ export default class WaDialog extends WebAwesomeElement {
   static shadowStyle = styles;
 
   private readonly localize = new LocalizeController(this);
+  private readonly hasSlotController = new HasSlotController(this, 'footer', 'header-actions', 'label');
   private originalTrigger: HTMLElement | null;
 
   @query('.dialog') dialog: HTMLDialogElement;
@@ -72,11 +74,8 @@ export default class WaDialog extends WebAwesomeElement {
    */
   @property({ reflect: true }) label = '';
 
-  /** Renders the dialog with a header. */
-  @property({ attribute: 'with-header', type: Boolean, reflect: true }) withHeader = false;
-
-  /** Renders the dialog with a footer. */
-  @property({ attribute: 'with-footer', type: Boolean, reflect: true }) withFooter = false;
+  /** Disables the header. This will also remove the default close button. */
+  @property({ attribute: 'without-header', type: Boolean, reflect: true }) withoutHeader = false;
 
   /** When enabled, the dialog will be closed when the user clicks outside of it. */
   @property({ attribute: 'light-dismiss', type: Boolean }) lightDismiss = false;
@@ -211,20 +210,25 @@ export default class WaDialog extends WebAwesomeElement {
   }
 
   render() {
+    const hasHeader =
+      !this.withoutHeader &&
+      (this.label.length > 0 || this.hasSlotController.test('label') || this.hasSlotController.test('header-actions'));
+    const hasFooter = this.hasSlotController.test('footer');
+
     return html`
       <dialog
         part="dialog"
         class=${classMap({
           dialog: true,
           'dialog--open': this.open,
-          'dialog--with-header': this.withHeader,
-          'dialog--with-footer': this.withFooter,
+          'dialog--has-header': hasHeader,
+          'dialog--has-footer': hasFooter,
         })}
         @cancel=${this.handleDialogCancel}
         @click=${this.handleDialogClick}
         @pointerdown=${this.handleDialogPointerDown}
       >
-        ${this.withHeader
+        ${hasHeader
           ? html`
               <header part="header" class="header">
                 <h2 part="title" class="title" id="title">
@@ -250,7 +254,7 @@ export default class WaDialog extends WebAwesomeElement {
 
         <div part="body" class="body"><slot></slot></div>
 
-        ${this.withFooter
+        ${hasFooter
           ? html`
               <footer part="footer" class="footer">
                 <slot name="footer"></slot>